Moved types to ./types
