+++ types/service.ts
@@ -1,4 +1,4 @@
-interface Service {
+export interface Service {
     service_id: number;
     name: string;
     price: string;
